fix: made suggested changes, move to common crate
