+++ crates/common/src/utils/predeploys.rs
@@ -1,22 +1,45 @@
 use alloy::primitives::{hex, Address};
 use alloy::{sol, sol_types::SolCall};
-use eyre::Result;
+use eyre::{eyre, Result};
+use once_cell::sync::Lazy;
+use serde::Deserialize;
+
+/// Pre-deploy contract data
+pub static PREDEPLOYS: Lazy<Vec<Predeploy>> = Lazy::new(|| {
+    const RAW: &str = include_str!("../data/predeploys.json");
+    serde_json::from_str(RAW).expect("invalid predeploys.json")
+});
+
+#[derive(Debug, Deserialize)]
+pub struct Predeploy {
+    pub address: Address,
+    pub constructor_input: String,
+}
 
 sol! {
-    /// EVM-predeploy manager function
+    /// EVM‐predeploy manager function
     function deployPredeployedContract(
         address contractAddress,
         bytes constructorInput
     ) external;
 }
 
-pub fn encode_predeploy_manager(addr: Address, ctor_hex: &str) -> Result<Vec<u8>> {
-    let ctor_bytes = hex::decode(ctor_hex.trim_start_matches("0x"))?;
+impl Predeploy {
+    pub fn encode_manager_call(&self) -> Result<Vec<u8>> {
+        let ctor_bytes =
+            hex::decode(self.constructor_input.trim_start_matches("0x")).map_err(|e| {
+                eyre!(
+                    "invalid hex in constructor_input for {}: {}",
+                    self.address,
+                    e
+                )
+            })?;
 
-    let call = deployPredeployedContractCall {
-        contractAddress: addr,
-        constructorInput: ctor_bytes.into(),
-    };
+        let call = deployPredeployedContractCall {
+            contractAddress: self.address,
+            constructorInput: ctor_bytes.into(),
+        };
 
-    Ok(call.abi_encode())
+        Ok(call.abi_encode())
+    }
 }
