+++ crates/types/Cargo.toml
@@ -21,7 +21,6 @@ clap.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 tracing.workspace = true
-once_cell.workspace = true
 
 [dev-dependencies]
 serde_json.workspace = true
