+++ crates/types/src/lib.rs
@@ -2,7 +2,6 @@ pub mod api;
 mod l2_tx_builder;
 mod log;
 pub mod numbers;
-pub mod predeploys;
 mod serde_helpers;
 mod show_details;
 pub mod traces;
@@ -11,7 +10,6 @@ mod transaction_order;
 pub use self::{
     l2_tx_builder::L2TxBuilder,
     log::LogLevel,
-    predeploys::{Predeploy, PREDEPLOYS},
     serde_helpers::Numeric,
     show_details::{ShowGasDetails, ShowStorageLogs, ShowVMDetails},
     transaction_order::{TransactionOrder, TransactionPriority},
